requirements.txt
# ReconXpert - Intelligent Attack Surface Mapper & Risk Reporter
# Requirements file for all dependencies

# Core networking and HTTP libraries
requests>=2.31.0
urllib3>=2.0.0

# DNS resolution and domain analysis
dnspython>=2.4.0
python-whois>=0.8.0

# HTML template rendering
jinja2>=3.1.0
markupsafe>=2.1.0

# Threat intelligence APIs
shodan>=1.30.0
censys>=2.2.0

# Data processing and analysis
python-dateutil>=2.8.0
pytz>=2023.3

# Security and cryptography (for API connections)
cryptography>=41.0.0
certifi>=2023.7.22

# Network scanning support (optional - for enhanced network analysis)
python-nmap>=0.7.1
scapy>=2.5.0

# JSON processing and validation
jsonschema>=4.19.0

# Additional utilities
tqdm>=4.66.0
colorama>=0.4.6
rich>=13.5.0

# Development and testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation (optional)
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Performance monitoring (optional)
psutil>=5.9.0
memory-profiler>=0.61.0
